+++ see/features.py
@@ -33,9 +33,9 @@ from .tools import compact
 
 PY_VERSION = sys.version_info
 
-if (PY_VERSION < (2, 6)) or ((3, 0) <= PY_VERSION < (3, 3)):  # pragma: nocover
+if (PY_VERSION < (2, 7)) or ((3, 0) <= PY_VERSION < (3, 4)):  # pragma: nocover
     sys.stderr.write('Warning: see() is not supported here. '
-                     'Please use Python 3.3+ or 2.6+.\n')
+                     'Please use Python 3.4+ or 2.7.\n')
 
 PY2 = PY_VERSION < (3, 0)
 PY3 = not PY2